+++ src/sero583/pacman/Vector2.java
@@ -1,43 +1,43 @@
 package sero583.pacman;
 
 public class Vector2 {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Vector2(int x, int y) {
+	public Vector2(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return this.x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return this.y;
 	}
 	
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
-	public Vector2 add(int x) {
+	public Vector2 add(double x) {
 		return this.add(x, 0);
 	}
 	
-	public Vector2 add(int x, int y) {
+	public Vector2 add(double x, double y) {
 		return new Vector2(this.x + x, this.y + y);
 	}
 	
-	public Vector2 substract(int x) {
+	public Vector2 substract(double x) {
 		return this.substract(x, 0);
 	}
 	
-	public Vector2 substract(int x, int y) {
+	public Vector2 substract(double x, double y) {
 		return new Vector2(this.x - x, this.y - y);
 	}
